requirements.txt
pywebview
requests
pyinstaller
protobuf
brotlipy
aiohttp